[PROG-14] Added grid support to SBSP (#13)
* added grid support to SBSP

* added reviewers suggestions, switched to full grid generation

* added reviewer suggestions

+++ src/core/components/transformation_component.h
@@ -2,13 +2,13 @@
 #include "src/core/common/point.h"
 
 struct TransformationComponent {
-    Point2f position;
+    Point2i position;
 
     TransformationComponent() = default;
 
     TransformationComponent(const TransformationComponent &) = default;
 
-    explicit TransformationComponent(const Point2f& position)
+    explicit TransformationComponent(const Point2i& position)
             : position(position) {}
 
 };
